Move the Daemon class to the service package
